[APPC-1247] Renamed wallet validation to poa wallet validation
