+++ app/src/main/java/com/asfoundation/wallet/poa_wallet_validation/ValidationLoadingFragment.kt
@@ -1,4 +1,4 @@
-package com.asfoundation.wallet.wallet_validation
+package com.asfoundation.wallet.poa_wallet_validation
 
 import android.content.Context
 import android.os.Bundle
