+++ app/src/main/java/com/asfoundation/wallet/poa_wallet_validation/WalletValidationBroadcastReceiver.kt
@@ -1,4 +1,4 @@
-package com.asfoundation.wallet.wallet_validation
+package com.asfoundation.wallet.poa_wallet_validation
 
 import android.app.NotificationManager
 import android.content.BroadcastReceiver
