+++ app/src/main/java/com/asfoundation/wallet/poa_wallet_validation/WalletValidationView.kt
@@ -1,4 +1,4 @@
-package com.asfoundation.wallet.wallet_validation
+package com.asfoundation.wallet.poa_wallet_validation
 
 interface WalletValidationView {
