+++ app/src/main/java/com/asfoundation/wallet/ui/onboarding/OnboardingView.kt
@@ -1,7 +1,7 @@
 package com.asfoundation.wallet.ui.onboarding
 
 import android.net.Uri
-import com.asfoundation.wallet.wallet_validation.WalletValidationStatus
+import com.asfoundation.wallet.poa_wallet_validation.WalletValidationStatus
 import io.reactivex.Observable
 
 interface OnboardingView {
